requirements.txt
flask
flask-cors
python-dotenv
openai
pinecone
langchain
langchain-openai
langchain-community
